[Refactor] domain 패키지 내부에 생성
