+++ src/main/java/com/inq/wishhair/wesharewishhair/user/service/UserService.java
@@ -1,9 +1,9 @@
 package com.inq.wishhair.wesharewishhair.user.service;
 
-import com.inq.wishhair.wesharewishhair.point.PointHistory;
-import com.inq.wishhair.wesharewishhair.point.repository.PointHistoryRepository;
-import com.inq.wishhair.wesharewishhair.user.User;
-import com.inq.wishhair.wesharewishhair.user.repository.UserRepository;
+import com.inq.wishhair.wesharewishhair.point.domain.PointHistory;
+import com.inq.wishhair.wesharewishhair.point.domain.PointHistoryRepository;
+import com.inq.wishhair.wesharewishhair.user.domain.User;
+import com.inq.wishhair.wesharewishhair.user.domain.UserRepository;
 import com.inq.wishhair.wesharewishhair.global.exception.ErrorCode;
 import com.inq.wishhair.wesharewishhair.global.exception.WishHairException;
 import lombok.RequiredArgsConstructor;
