+++ src/main/java/com/inq/wishhair/wesharewishhair/web/review/ReviewController.java
@@ -1,7 +1,7 @@
 package com.inq.wishhair.wesharewishhair.web.review;
 
 import com.inq.wishhair.wesharewishhair.auth.config.resolver.ExtractPayload;
-import com.inq.wishhair.wesharewishhair.review.Review;
+import com.inq.wishhair.wesharewishhair.review.domain.Review;
 import com.inq.wishhair.wesharewishhair.review.service.ReviewService;
 import com.inq.wishhair.wesharewishhair.web.review.dto.request.ReviewRequest;
 import com.inq.wishhair.wesharewishhair.web.review.dto.response.ReviewsResponse;
