+++ src/main/java/com/inq/wishhair/wesharewishhair/wishlist/service/WishListService.java
@@ -1,9 +1,9 @@
 package com.inq.wishhair.wesharewishhair.wishlist.service;
 
-import com.inq.wishhair.wesharewishhair.hairstyle.HairStyle;
-import com.inq.wishhair.wesharewishhair.hairstyle.repository.HairStyleRepository;
-import com.inq.wishhair.wesharewishhair.user.User;
-import com.inq.wishhair.wesharewishhair.user.repository.UserRepository;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyleRepository;
+import com.inq.wishhair.wesharewishhair.user.domain.User;
+import com.inq.wishhair.wesharewishhair.user.domain.UserRepository;
 import com.inq.wishhair.wesharewishhair.wishlist.WishList;
 import com.inq.wishhair.wesharewishhair.wishlist.repository.WishListRepository;
 import com.inq.wishhair.wesharewishhair.global.exception.ErrorCode;
