+++ src/test/java/com/inq/wishhair/wesharewishhair/fixture/UserFixture.java
@@ -1,6 +1,6 @@
 package com.inq.wishhair.wesharewishhair.fixture;
 
-import com.inq.wishhair.wesharewishhair.user.User;
+import com.inq.wishhair.wesharewishhair.user.domain.User;
 import com.inq.wishhair.wesharewishhair.user.enums.Sex;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
