+++ src/test/java/com/inq/wishhair/wesharewishhair/global/base/ServiceTest.java
@@ -2,7 +2,7 @@ package com.inq.wishhair.wesharewishhair.global.base;
 
 import com.inq.wishhair.wesharewishhair.WeShareWishHairApplication;
 import com.inq.wishhair.wesharewishhair.global.testrepository.PointHistoryTestRepository;
-import com.inq.wishhair.wesharewishhair.hairstyle.repository.HairStyleRepository;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyleRepository;
 import com.inq.wishhair.wesharewishhair.domain.login.LoginService;
 import com.inq.wishhair.wesharewishhair.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
