+++ src/test/java/com/inq/wishhair/wesharewishhair/user/domain/UserTest.java
@@ -1,6 +1,5 @@
 package com.inq.wishhair.wesharewishhair.user.domain;
 
-import com.inq.wishhair.wesharewishhair.user.User;
 import com.inq.wishhair.wesharewishhair.fixture.UserFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
